pyproject.toml
[project]
name = "speech-to-speech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask[async]>=3.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "mcp>=1.8.1",
    "mcp-use>=1.2.12",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "wave>=0.0.2",
    "werkzeug>=3.1.3",
]
